pyproject.toml
[project]
name = "iot-dashboard"
version = "0.1.0"
description = "IoT Dashboard for monitoring and controlling IoT services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.16.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "streamlit>=1.53.0",
    "streamlit-autorefresh>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "mongomock>=4.2.0",
    "responses>=0.25.3",
]

[tool.hatch.build.targets.wheel]
packages = ["dashboard"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
